Move MuxAsyncIterator to util/async.ts
